backend/requirements.txt
# =============================================================================
# 3D Development Platform - Backend Dependencies
# =============================================================================

# --- Web Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# --- Database ---
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
geoalchemy2==0.15.2

# --- Cache & Task Queue ---
redis==5.2.1
celery[redis]==5.4.0

# --- Authentication ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# --- Document Processing ---
PyMuPDF==1.25.1
opencv-python-headless==4.10.0.84
ezdxf==1.3.4
pandas==2.2.3
openpyxl==3.1.5
Pillow==11.1.0
pytesseract==0.3.13

# --- AI & ML ---
anthropic==0.42.0
transformers==4.47.1

# --- 3D Generation ---
mapbox-earcut==2.0.0
trimesh==4.5.3
pygltflib==1.16.3
numpy==1.26.4
scipy==1.14.1

# --- Object Storage ---
boto3==1.35.81

# --- Validation & Serialization ---
pydantic[email]==2.10.3
pydantic-settings==2.7.0

# --- Geospatial ---
shapely==2.0.6
pyproj==3.7.0
geojson==3.2.0

# --- PDF Generation ---
fpdf2==2.8.1

# --- Monitoring ---
sentry-sdk[fastapi]==2.19.2

# --- Testing ---
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
factory-boy==3.3.1

# --- Dev Tools ---
black==24.10.0
ruff==0.8.4
mypy==1.13.0
